+++ src/interface/src/app/shared/form-response-box/form-message-box.component.scss
