+++ src/interface/src/app/shared/form-response-box/form-message-box.component.spec.ts
@@ -1,6 +1,6 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 
-import { FormMessageBoxComponent } from './form-response-box.component';
+import { FormMessageBoxComponent } from './form-message-box.component';
 
 describe('FormMessageBoxComponent', () => {
   let component: FormMessageBoxComponent;
