+++ src/interface/src/app/shared/shared.module.ts
@@ -12,7 +12,7 @@ import { CurrencyInKPipe } from '../pipes/currency-in-k.pipe';
 import { TypeSafeMatCellDef } from './type-safe-mat-cell/type-safe-mat-cell-def.directive';
 import { FieldAlertComponent } from './field-alert/field-alert.component';
 import { CreditsBlurbComponent } from './credits-blurb/credits-blurb.component';
-import { FormMessageBoxComponent } from './form-response-box/form-response-box.component';
+import { FormMessageBoxComponent } from './form-response-box/form-message-box.component';
 
 @NgModule({
   declarations: [
